wip move table names to their own file. wip table name enum translation
