+++ mappings/table_name_enum_translation.go
@@ -5,21 +5,27 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/cmsgov/mint-app/pkg/graph/model"
 	"github.com/cmsgov/mint-app/pkg/models"
 )
 
 //go:embed translation/table_name.json
 var tableNameEnumJSON []byte
+var TableNamesTranslationMap = tableNamesTranslationMapFunc()
 
 // TableNamesTranslationMap provides a map of TableName to enum translation
-func TableNamesTranslationMap() (map[models.TableName]model.EnumTranslation, error) {
-	var translation map[models.TableName]model.EnumTranslation
+func tableNamesTranslationMapFunc() map[models.TableName]models.EnumTranslation {
+	var translation map[models.TableName]models.EnumTranslation
 	err := json.Unmarshal(tableNameEnumJSON, &translation)
 	if err != nil {
-		fmt.Println("Error unmarshalling JSON:", err)
-		return nil, err
-	}
-	return translation, nil
+		err := fmt.Errorf("error unmarshalling JSON for tableNameTranslationMap err: %w", err)
+		panic(err)
 
+	}
+	return translation
+}
+
+// TranslateTableName will return a Translation for a table name
+func TranslateTableName(tableName models.TableName) (models.EnumTranslation, bool) {
+	translation, wasFound := TableNamesTranslationMap[tableName]
+	return translation, wasFound
 }
