+++ mappings/table_name_enum_translation_test.go
@@ -11,8 +11,7 @@ import (
 func TestTableNamesTranslationMap(t *testing.T) {
 	assert := assert.New(t)
 
-	translationMap, err := TableNamesTranslationMap()
-	assert.NoError(err)
+	translationMap := TableNamesTranslationMap
 	assert.NotNil(translationMap)
 
 	// TODO: iterate over all table names and assert there is a translation
